backend/requirements.txt
fastapi>=0.104.1
uvicorn>=0.24.0
openai-agents==0.8.1
qdrant-client>=1.8.0
tiktoken>=0.5.0
python-dotenv>=1.0.0